Test cases have been created.

+++ app/src/main/java/com/example/achievav2/AnalyzerActivity.kt
@@ -22,8 +22,8 @@ class AnalyzerActivity : AppCompatActivity() {
             if (!textInputModerate.text.toString().equals("") && !textInputVigorous.text.toString().equals(""))
             {
 
-                val inputWeeklyModerateActivity = textInputModerate.text.toString().toFloat()
-                val inputWeeklyVigorousActivity = textInputVigorous.text.toString().toFloat()
+                val inputWeeklyModerateActivity = textInputModerate.text.toString().toDouble()
+                val inputWeeklyVigorousActivity = textInputVigorous.text.toString().toDouble()
 
 
                 //There are 4 scenarios:
